feat: unify workflow editor — reuse canvas in Add New Workflow view
Extract WorkflowEditorCanvas from the definition page into a shared
component. Add New Workflow now uses the same diagram, toolbar, step
editor, and YAML preview as the Workflow Editor. Starts with a 2-step
template (human draft + AI review) so new users have a working example.
Header collects Workflow ID, Owner, and Description before creating.
