+++ packages/platform-ui/src/app/(app)/[handle]/workflows/[name]/definitions/[version]/page.tsx
@@ -1,29 +1,16 @@
 'use client';
 
-import { useState, useCallback, useEffect, useRef, useMemo } from 'react';
+import { useState, useCallback } from 'react';
 import Link from 'next/link';
 import { useParams, useRouter } from 'next/navigation';
-import { X, Save, User, Bot, Terminal } from 'lucide-react';
-import { stringify as yamlStringify, parse as yamlParse } from 'yaml';
+import { Save } from 'lucide-react';
 import { useWorkflowDefinitions } from '@/hooks/use-workflow-definitions';
-import { useAuth } from '@/contexts/auth-context';
-import { usePlugins } from '@/hooks/use-plugins';
-import { WorkflowDiagram } from '@/components/workflows/workflow-diagram';
+import { WorkflowEditorCanvas } from '@/components/workflows/workflow-editor-canvas';
 import { saveWorkflowDefinition } from '@/app/actions/definitions';
-import { getWorkflowSecretKeys } from '@/app/actions/workflow-secrets';
 import { VersionLabel } from '@/components/ui/version-label';
 import { cn } from '@/lib/utils';
 import { routes } from '@/lib/routes';
 import type { WorkflowDefinition, WorkflowStep } from '@mediforce/platform-core';
-
-const EXECUTOR_TYPES = ['human', 'agent', 'script'] as const;
-const AUTONOMY_LEVELS = [
-  { value: 'L0', label: 'L0 — Manual only' },
-  { value: 'L1', label: 'L1 — Human review' },
-  { value: 'L2', label: 'L2 — Auto if confident' },
-  { value: 'L3', label: 'L3 — Auto + fallback' },
-  { value: 'L4', label: 'L4 — Full autonomy' },
-] as const;
 
 type SaveState =
   | { status: 'idle' }
@@ -41,254 +28,30 @@ export default function WorkflowDefinitionVersionPage() {
   const definition = definitions.find((def) => def.version === versionNumber) ?? null;
 
   const [editedTitle, setEditedTitle] = useState('');
-  const [editedSteps, setEditedSteps] = useState<WorkflowStep[]>([]);
-  const [editedTransitions, setEditedTransitions] = useState<WorkflowDefinition['transitions']>([]);
-  const [selectedStepId, setSelectedStepId] = useState<string | null>(null);
   const [saveState, setSaveState] = useState<SaveState>({ status: 'idle' });
-  const [addingStep, setAddingStep] = useState(false);
-  const [pendingStepType, setPendingStepType] = useState<WorkflowStep['type'] | null>(null);
-  const [editHistory, setEditHistory] = useState<Array<{ steps: WorkflowStep[]; transitions: WorkflowDefinition['transitions'] }>>([]);
 
-  // Initialize editable state from definition whenever it loads or changes version
-  useEffect(() => {
-    if (!definition) return;
-    setEditedSteps(structuredClone(definition.steps));
-    setEditedTransitions(structuredClone(definition.transitions));
-    setEditedTitle('');
-    setEditHistory([]);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [definition?.version]);
-
-  const currentSteps = editedSteps.length > 0 ? editedSteps : (definition?.steps ?? []);
-  const currentTransitions = editedSteps.length > 0 ? editedTransitions : (definition?.transitions ?? []);
-
-  const hasChanges = useMemo(() => {
-    if (!definition || editedSteps.length === 0) return false;
-    return (
-      JSON.stringify(editedSteps) !== JSON.stringify(definition.steps) ||
-      JSON.stringify(editedTransitions) !== JSON.stringify(definition.transitions)
-    );
-  }, [editedSteps, editedTransitions, definition]);
-  const selectedStep = currentSteps.find((s) => s.id === selectedStepId) ?? null;
-
-  const canMoveSelectedUp = (() => {
-    if (!selectedStepId) return false;
-    const incoming = currentTransitions.filter((t) => t.to === selectedStepId);
-    if (incoming.length !== 1) return false;
-    const pred = incoming[0].from;
-    return currentTransitions.filter((t) => t.from === pred).length === 1;
-  })();
-
-  const canMoveSelectedDown = (() => {
-    if (!selectedStepId) return false;
-    const outgoing = currentTransitions.filter((t) => t.from === selectedStepId);
-    if (outgoing.length !== 1) return false;
-    const succ = outgoing[0].to;
-    return currentTransitions.filter((t) => t.to === succ).length === 1;
-  })();
-
-  const saveSnapshot = useCallback(() => {
-    setEditHistory((prev) => [...prev, { steps: editedSteps, transitions: editedTransitions }]);
-  }, [editedSteps, editedTransitions]);
-
-  const undoEdit = useCallback(() => {
-    setEditHistory((prev) => {
-      if (prev.length === 0) return prev;
-      const snapshot = prev[prev.length - 1];
-      setEditedSteps(snapshot.steps);
-      setEditedTransitions(snapshot.transitions);
-      return prev.slice(0, -1);
-    });
-  }, []);
-
-  const discardChanges = useCallback(() => {
-    if (!definition) return;
-    setEditedSteps(structuredClone(definition.steps));
-    setEditedTransitions(structuredClone(definition.transitions));
-    setEditedTitle('');
-    setSaveState({ status: 'idle' });
-    setEditHistory([]);
-  }, [definition]);
-
-  const updateStep = useCallback((stepId: string, patch: Partial<WorkflowStep>) => {
-    setEditedSteps((prev) =>
-      prev.map((s) => (s.id === stepId ? { ...s, ...patch } : s)),
-    );
-    // When step ID changes, update all references: transitions + verdict targets in other steps
-    if (patch.id && patch.id !== stepId) {
-      const newId = patch.id;
-      setEditedTransitions((prev) =>
-        prev.map((t) => ({
-          from: t.from === stepId ? newId : t.from,
-          to: t.to === stepId ? newId : t.to,
-          ...(t.when ? { when: t.when } : {}),
-        })),
-      );
-      // Update verdict targets in other steps that point to the renamed step
-      setEditedSteps((prev) =>
-        prev.map((s) => {
-          if (!s.verdicts) return s;
-          const hasRef = Object.values(s.verdicts).some((v) => v.target === stepId);
-          if (!hasRef) return s;
-          const updatedVerdicts: Record<string, { target: string }> = {};
-          for (const [name, v] of Object.entries(s.verdicts)) {
-            updatedVerdicts[name] = { target: v.target === stepId ? newId : v.target };
-          }
-          return { ...s, verdicts: updatedVerdicts };
-        }),
-      );
-      if (selectedStepId === stepId) setSelectedStepId(newId);
-    }
-  }, [selectedStepId]);
-
-  // Ctrl+Z keyboard shortcut for undo
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'z' && (e.ctrlKey || e.metaKey) && !e.shiftKey) {
-        e.preventDefault();
-        undoEdit();
-      }
-    };
-    document.addEventListener('keydown', handleKeyDown);
-    return () => document.removeEventListener('keydown', handleKeyDown);
-  }, [undoEdit]);
-
-  const addStep = useCallback((type: WorkflowStep['type'], executor: WorkflowStep['executor']) => {
-    saveSnapshot();
-    // Insert after selected step, or after the last step if nothing is selected
-    const afterId = selectedStepId ?? editedSteps[editedSteps.length - 1]?.id;
-    if (!afterId) return;
-    const stepNum = editedSteps.length + 1;
-    const newId = `new-step-${stepNum}`;
-    setEditedSteps((prev) => {
-      const idx = prev.findIndex((s) => s.id === afterId);
-      if (idx === -1) return prev;
-      const newStep: WorkflowStep = {
-        id: newId,
-        name: `New Step ${stepNum}`,
-        type,
-        executor,
-        ...(executor === 'agent' ? { plugin: 'opencode-agent', autonomyLevel: 'L2' } : {}),
-        ...(executor === 'script' ? { plugin: 'script-container' } : {}),
-      };
-      const next = [...prev];
-      next.splice(idx + 1, 0, newStep);
-      return next;
-    });
-    setEditedTransitions((prev) => {
-      const existingEdge = prev.find((t) => t.from === afterId);
-      const without = existingEdge ? prev.filter((t) => t !== existingEdge) : prev;
-      if (type === 'terminal') {
-        return [...without, { from: afterId, to: newId }];
-      }
-      return existingEdge
-        ? [...without, { from: afterId, to: newId }, { from: newId, to: existingEdge.to }]
-        : [...without, { from: afterId, to: newId }];
-    });
-    setSelectedStepId(newId);
-    setAddingStep(false);
-    setPendingStepType(null);
-  }, [selectedStepId, editedSteps, saveSnapshot]);
-
-  const removeStep = useCallback((stepId: string) => {
-    saveSnapshot();
-    setEditedSteps((prev) => prev.filter((s) => s.id !== stepId));
-    // Rewire transitions: if A→removed→B, create A→B
-    setEditedTransitions((prev) => {
-      const incoming = prev.filter((t) => t.to === stepId);
-      const outgoing = prev.filter((t) => t.from === stepId);
-      const unrelated = prev.filter((t) => t.from !== stepId && t.to !== stepId);
-      const rewired = incoming.flatMap((inc) =>
-        outgoing.map((out) => ({ from: inc.from, to: out.to })),
-      );
-      return [...unrelated, ...rewired];
-    });
-    if (selectedStepId === stepId) setSelectedStepId(null);
-  }, [selectedStepId, saveSnapshot]);
-
-  const moveStep = useCallback((stepId: string, direction: 'up' | 'down') => {
-    saveSnapshot();
-    setEditedTransitions((prev) => {
-      if (direction === 'up') {
-        const incoming = prev.filter((t) => t.to === stepId);
-        if (incoming.length !== 1) return prev;
-        const pred = incoming[0].from;
-        const predIncoming = prev.filter((t) => t.to === pred);
-        const predOutgoing = prev.filter((t) => t.from === pred);
-        if (predOutgoing.length !== 1) return prev;
-        const stepOutgoing = prev.filter((t) => t.from === stepId);
-        // Before: [...→pred→stepId→succ]  After: [...→stepId→pred→succ]
-        const toRemove = new Set([
-          ...predIncoming.map((t) => `${t.from}|${t.to}`),
-          `${pred}|${stepId}`,
-          ...stepOutgoing.map((t) => `${t.from}|${t.to}`),
-        ]);
-        return [
-          ...prev.filter((t) => !toRemove.has(`${t.from}|${t.to}`)),
-          ...predIncoming.map((t) => ({ from: t.from, to: stepId })),
-          { from: stepId, to: pred },
-          ...stepOutgoing.map((t) => ({ from: pred, to: t.to })),
-        ];
-      } else {
-        const outgoing = prev.filter((t) => t.from === stepId);
-        if (outgoing.length !== 1) return prev;
-        const succ = outgoing[0].to;
-        const succIncoming = prev.filter((t) => t.to === succ);
-        if (succIncoming.length !== 1) return prev;
-        const succOutgoing = prev.filter((t) => t.from === succ);
-        const stepIncoming = prev.filter((t) => t.to === stepId);
-        // Before: [pred→stepId→succ→...]  After: [pred→succ→stepId→...]
-        const toRemove = new Set([
-          ...stepIncoming.map((t) => `${t.from}|${t.to}`),
-          `${stepId}|${succ}`,
-          ...succOutgoing.map((t) => `${t.from}|${t.to}`),
-        ]);
-        return [
-          ...prev.filter((t) => !toRemove.has(`${t.from}|${t.to}`)),
-          ...stepIncoming.map((t) => ({ from: t.from, to: succ })),
-          { from: succ, to: stepId },
-          ...succOutgoing.map((t) => ({ from: stepId, to: t.to })),
-        ];
-      }
-    });
-    // Keep steps array order in sync
-    setEditedSteps((prev) => {
-      const idx = prev.findIndex((s) => s.id === stepId);
-      if (idx === -1) return prev;
-      const next = [...prev];
-      if (direction === 'up' && idx > 0) {
-        [next[idx - 1], next[idx]] = [next[idx], next[idx - 1]];
-      } else if (direction === 'down' && idx < next.length - 1) {
-        [next[idx], next[idx + 1]] = [next[idx + 1], next[idx]];
-      }
-      return next;
-    });
-  }, [saveSnapshot]);
-
-  const handleSave = useCallback(async () => {
+  const handleSave = useCallback(async (
+    steps: WorkflowStep[],
+    transitions: WorkflowDefinition['transitions'],
+  ) => {
     if (!definition) return;
 
-    // Validate: agent/script steps must have a plugin set
-    const missingPlugin = editedSteps.filter(
+    const missingPlugin = steps.filter(
       (s) => s.type !== 'terminal' && (s.executor === 'agent' || s.executor === 'script') && !s.plugin,
     );
     if (missingPlugin.length > 0) {
-      const names = missingPlugin.map((s) => `"${s.name}"`).join(', ');
-      setSaveState({ status: 'error', message: `Plugin required for agent/script steps: ${names}` });
+      setSaveState({ status: 'error', message: `Plugin required for agent/script steps: ${missingPlugin.map((s) => `"${s.name}"`).join(', ')}` });
       return;
     }
 
-    // Validate: steps must have non-empty IDs
-    const emptyIds = editedSteps.filter((s) => !s.id);
+    const emptyIds = steps.filter((s) => !s.id);
     if (emptyIds.length > 0) {
-      const names = emptyIds.map((s) => `"${s.name}"`).join(', ');
-      setSaveState({ status: 'error', message: `Step ID is empty for: ${names}` });
+      setSaveState({ status: 'error', message: `Step ID is empty for: ${emptyIds.map((s) => `"${s.name}"`).join(', ')}` });
       return;
     }
 
-    // Validate: no duplicate step IDs
     const idCounts = new Map<string, number>();
-    for (const s of editedSteps) idCounts.set(s.id, (idCounts.get(s.id) ?? 0) + 1);
+    for (const s of steps) idCounts.set(s.id, (idCounts.get(s.id) ?? 0) + 1);
     const dupes = [...idCounts.entries()].filter(([, count]) => count > 1).map(([id]) => id);
     if (dupes.length > 0) {
       setSaveState({ status: 'error', message: `Duplicate step IDs: ${dupes.join(', ')}` });
@@ -297,13 +60,12 @@ export default function WorkflowDefinitionVersionPage() {
 
     setSaveState({ status: 'saving' });
 
-    // Merge explicit transitions with verdict-based transitions
-    const transitions = [...editedTransitions];
-    for (const step of editedSteps) {
+    const mergedTransitions = [...transitions];
+    for (const step of steps) {
       if (step.type === 'review' && step.verdicts) {
         for (const verdict of Object.values(step.verdicts)) {
-          if (verdict.target && !transitions.some((t) => t.from === step.id && t.to === verdict.target)) {
-            transitions.push({ from: step.id, to: verdict.target });
+          if (verdict.target && !mergedTransitions.some((t) => t.from === step.id && t.to === verdict.target)) {
+            mergedTransitions.push({ from: step.id, to: verdict.target });
           }
         }
       }
@@ -313,8 +75,8 @@ export default function WorkflowDefinitionVersionPage() {
       name: definition.name,
       title: editedTitle.trim() || undefined,
       description: definition.description,
-      steps: editedSteps,
-      transitions,
+      steps,
+      transitions: mergedTransitions,
       triggers: definition.triggers,
       roles: definition.roles,
       env: definition.env,
@@ -332,12 +94,59 @@ export default function WorkflowDefinitionVersionPage() {
     } else {
       setSaveState({ status: 'error', message: result.error });
     }
-  }, [definition, editedTitle, editedSteps, editedTransitions, name, router]);
+  }, [definition, editedTitle, name, handle, router]);
 
-  // Build a WorkflowDefinition from edited steps for the diagram
-  const diagramDefinition: WorkflowDefinition | null = definition
-    ? { ...definition, steps: currentSteps, transitions: currentTransitions }
-    : null;
+  const renderSavePanel = useCallback((
+    steps: WorkflowStep[],
+    transitions: WorkflowDefinition['transitions'],
+    onDiscard: () => void,
+  ) => {
+    if (!definition) return null;
+    const hasChanges =
+      JSON.stringify(steps) !== JSON.stringify(definition.steps) ||
+      JSON.stringify(transitions) !== JSON.stringify(definition.transitions);
+    if (!hasChanges) return null;
+
+    return (
+      <div className="space-y-3">
+        <input
+          value={editedTitle}
+          onChange={(e) => setEditedTitle(e.target.value)}
+          placeholder="Version title (required) — e.g. &quot;Added automated review step&quot;"
+          className="w-full text-sm border rounded-md px-2.5 py-1 bg-background focus:outline-none focus:ring-1 focus:ring-primary"
+        />
+        <div className="flex items-center gap-2 flex-wrap">
+          <button
+            onClick={() => { setEditedTitle(''); setSaveState({ status: 'idle' }); onDiscard(); }}
+            className="inline-flex items-center gap-1.5 rounded-md border px-3 py-1.5 text-sm font-medium text-muted-foreground hover:text-foreground hover:bg-muted transition-colors whitespace-nowrap"
+          >
+            Discard changes
+          </button>
+          <button
+            onClick={() => handleSave(steps, transitions)}
+            disabled={saveState.status === 'saving' || !editedTitle.trim()}
+            className={cn(
+              'inline-flex items-center gap-1.5 rounded-md bg-primary px-3 py-1.5 text-sm font-medium text-primary-foreground hover:bg-primary/90 transition-colors whitespace-nowrap',
+              (saveState.status === 'saving' || !editedTitle.trim()) && 'opacity-50 cursor-not-allowed',
+            )}
+          >
+            <Save className="h-3.5 w-3.5" />
+            {saveState.status === 'saving' ? 'Saving...' : 'Save new version'}
+          </button>
+          {saveState.status === 'saved' && (
+            <span className="inline-flex items-center gap-1.5 rounded-md bg-green-50 border border-green-200 px-3 py-1.5 text-sm font-medium text-green-700 dark:bg-green-900/20 dark:border-green-800 dark:text-green-400">
+              Saved as v{saveState.version}
+            </span>
+          )}
+          {saveState.status === 'error' && (
+            <span className="inline-flex items-center gap-1.5 rounded-md bg-red-50 border border-red-200 px-3 py-1.5 text-sm text-red-700 dark:bg-red-900/20 dark:border-red-800 dark:text-red-400">
+              {saveState.message}
+            </span>
+          )}
+        </div>
+      </div>
+    );
+  }, [definition, editedTitle, saveState, handleSave]);
 
   if (loading) {
     return (
@@ -358,979 +167,39 @@ export default function WorkflowDefinitionVersionPage() {
     );
   }
 
-  if (definition === null || diagramDefinition === null) return null;
+  if (definition === null) return null;
 
   return (
     <div className="flex flex-1 flex-col relative">
       {/* Header */}
       <div className="border-b px-6 py-4 sticky top-0 z-30 bg-background">
-        <div className="flex items-start justify-between gap-4">
-          <div className="space-y-0.5">
-            <div className="flex items-baseline gap-2">
-              <span className="text-xs text-muted-foreground w-24 shrink-0">Workflow ID</span>
-              <span className="text-sm font-mono">{decodedName}</span>
-            </div>
-            <div className="flex items-baseline gap-2">
-              <span className="text-xs text-muted-foreground w-24 shrink-0">Version</span>
-              <VersionLabel version={definition.version} title={definition.title} className="text-sm" />
-            </div>
-            {definition.description && (
-              <div className="flex items-baseline gap-2">
-                <span className="text-xs text-muted-foreground w-24 shrink-0">Description</span>
-                <span className="text-sm text-muted-foreground">{definition.description}</span>
-              </div>
-            )}
+        <div className="space-y-0.5">
+          <div className="flex items-baseline gap-2">
+            <span className="text-xs text-muted-foreground w-24 shrink-0">Workflow ID</span>
+            <span className="text-sm font-mono">{decodedName}</span>
           </div>
-
+          <div className="flex items-baseline gap-2">
+            <span className="text-xs text-muted-foreground w-24 shrink-0">Version</span>
+            <VersionLabel version={definition.version} title={definition.title} className="text-sm" />
+          </div>
+          {definition.description && (
+            <div className="flex items-baseline gap-2">
+              <span className="text-xs text-muted-foreground w-24 shrink-0">Description</span>
+              <span className="text-sm text-muted-foreground">{definition.description}</span>
+            </div>
+          )}
         </div>
       </div>
 
-      {/* Content: diagram + side panel */}
-      <div className="flex flex-1 min-h-0">
-        {/* Diagram */}
-        <div className="flex-1 flex flex-col pr-0">
-          {/* Editing toolbar */}
-          <div className="border-b px-4 py-2 flex items-center gap-1.5 bg-muted/30 shrink-0 flex-wrap">
-              {/* Add Step */}
-              <div className="relative">
-                <button
-                  onClick={() => { setAddingStep(!addingStep); setPendingStepType(null); }}
-                  className="inline-flex items-center gap-1.5 rounded-md bg-primary px-3 py-1.5 text-sm font-medium text-primary-foreground hover:bg-primary/90 transition-colors"
-                >
-                  + Add Step
-                </button>
-                {addingStep && (
-                  <div className="absolute top-full left-0 mt-1.5 bg-background border rounded-xl shadow-xl p-3 z-50 w-64 space-y-3">
-                    <div>
-                      <p className="text-[10px] font-semibold uppercase tracking-widest text-muted-foreground mb-2">Step type</p>
-                      <div className="grid grid-cols-2 gap-1.5">
-                        {([
-                          { type: 'creation', label: 'Creation', color: 'text-blue-600 dark:text-blue-400', bg: pendingStepType === 'creation' ? 'bg-blue-100 dark:bg-blue-900/50 ring-1 ring-blue-400' : 'hover:bg-muted' },
-                          { type: 'review', label: 'Review', color: 'text-amber-600 dark:text-amber-400', bg: pendingStepType === 'review' ? 'bg-amber-100 dark:bg-amber-900/50 ring-1 ring-amber-400' : 'hover:bg-muted' },
-                          { type: 'decision', label: 'Decision', color: 'text-purple-600 dark:text-purple-400', bg: pendingStepType === 'decision' ? 'bg-purple-100 dark:bg-purple-900/50 ring-1 ring-purple-400' : 'hover:bg-muted' },
-                          { type: 'terminal', label: 'End', color: 'text-emerald-600 dark:text-emerald-400', bg: 'hover:bg-muted' },
-                        ] as const).map((opt) => (
-                          <button
-                            key={opt.type}
-                            onClick={() => {
-                              if (opt.type === 'terminal') { addStep('terminal', 'human'); }
-                              else { setPendingStepType(opt.type); }
-                            }}
-                            className={cn('rounded-lg px-3 py-2 text-xs font-semibold transition-all text-left', opt.color, opt.bg)}
-                          >
-                            {opt.label}
-                          </button>
-                        ))}
-                      </div>
-                    </div>
-                    {pendingStepType && (
-                      <div>
-                        <p className="text-[10px] font-semibold uppercase tracking-widest text-muted-foreground mb-2">Executor</p>
-                        <div className="flex gap-1.5">
-                          {(['creation', 'review', 'decision'].includes(pendingStepType) ? (
-                            pendingStepType === 'creation'
-                              ? (['human', 'agent', 'script'] as const)
-                              : (['human', 'agent'] as const)
-                          ) : []).map((executor) => (
-                            <button
-                              key={executor}
-                              onClick={() => addStep(pendingStepType, executor)}
-                              className="flex-1 rounded-lg py-1.5 text-xs font-semibold hover:bg-muted transition-all capitalize border"
-                            >
-                              {executor}
-                            </button>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                )}
-              </div>
-
-              <div className="w-px h-4 bg-border mx-0.5" />
-
-              {/* Move Up */}
-              <button
-                onClick={() => selectedStepId && moveStep(selectedStepId, 'up')}
-                disabled={!canMoveSelectedUp}
-                className={cn(
-                  'inline-flex items-center gap-1.5 rounded-md px-3 py-1.5 text-sm font-medium border transition-colors',
-                  canMoveSelectedUp
-                    ? 'hover:bg-muted text-foreground'
-                    : 'opacity-40 cursor-not-allowed text-muted-foreground',
-                )}
-              >
-                ↑ Move Up
-              </button>
-
-              {/* Move Down */}
-              <button
-                onClick={() => selectedStepId && moveStep(selectedStepId, 'down')}
-                disabled={!canMoveSelectedDown}
-                className={cn(
-                  'inline-flex items-center gap-1.5 rounded-md px-3 py-1.5 text-sm font-medium border transition-colors',
-                  canMoveSelectedDown
-                    ? 'hover:bg-muted text-foreground'
-                    : 'opacity-40 cursor-not-allowed text-muted-foreground',
-                )}
-              >
-                ↓ Move Down
-              </button>
-
-              {/* Undo */}
-              <button
-                onClick={undoEdit}
-                disabled={editHistory.length === 0}
-                title="Undo last change (Ctrl+Z)"
-                className={cn(
-                  'inline-flex items-center gap-1.5 rounded-md px-3 py-1.5 text-sm font-medium border transition-colors',
-                  editHistory.length > 0
-                    ? 'hover:bg-muted text-foreground'
-                    : 'opacity-40 cursor-not-allowed text-muted-foreground',
-                )}
-              >
-                ↩ Undo
-              </button>
-
-              <div className="w-px h-4 bg-border mx-0.5" />
-
-              {/* Remove Step */}
-              <button
-                onClick={() => selectedStepId && removeStep(selectedStepId)}
-                disabled={!selectedStepId}
-                className={cn(
-                  'inline-flex items-center gap-1.5 rounded-md px-3 py-1.5 text-sm font-medium border transition-colors',
-                  selectedStepId
-                    ? 'hover:bg-red-50 hover:border-red-200 hover:text-red-600 text-foreground dark:hover:bg-red-900/20 dark:hover:text-red-400'
-                    : 'opacity-40 cursor-not-allowed text-muted-foreground',
-                )}
-              >
-                Remove Step
-              </button>
-
-              {selectedStepId && (
-                <span className="ml-auto text-xs text-muted-foreground">
-                  Selected: <span className="font-mono">{selectedStepId}</span>
-                </span>
-              )}
-            </div>
-
-          <div className="flex-1 p-6 pt-4">
-          <WorkflowDiagram
-            definition={diagramDefinition}
-            className="border-0"
-            onNodeClick={(stepId) => setSelectedStepId(stepId === selectedStepId ? null : stepId)}
-            selectedStepId={selectedStepId}
-          />
-          </div>
-        </div>
-
-        {/* Side panel — always visible; shows step details when a step is selected, YAML otherwise */}
-        <div className="w-1/2 shrink-0 border-l bg-background overflow-y-auto">
-          <div className="p-4 space-y-4">
-            {selectedStep ? (
-              <>
-                <div className="flex items-center justify-between">
-                  <h2 className="text-sm font-semibold">Edit step</h2>
-                  <button
-                    onClick={() => setSelectedStepId(null)}
-                    className="rounded-md p-1 text-muted-foreground hover:text-foreground hover:bg-muted transition-colors"
-                  >
-                    <X className="h-4 w-4" />
-                  </button>
-                </div>
-                <StepEditor
-                  step={selectedStep}
-                  allSteps={currentSteps}
-                  onChange={(patch) => updateStep(selectedStep.id, patch)}
-                />
-              </>
-            ) : (
-              <>
-                <div className="flex items-center justify-between">
-                  <h2 className="text-sm font-semibold">YAML</h2>
-                </div>
-                <pre className="text-[11px] font-mono bg-muted/30 rounded-lg p-4 overflow-x-auto overflow-y-auto leading-relaxed">
-                  {yamlStringify(
-                    { ...diagramDefinition, version: undefined, createdAt: undefined },
-                    { indent: 2 },
-                  )}
-                </pre>
-                {hasChanges && (
-                  <div className="border-t pt-4 space-y-3">
-                    <input
-                      value={editedTitle}
-                      onChange={(e) => setEditedTitle(e.target.value)}
-                      placeholder="Version title (required) — e.g. &quot;Added automated review step&quot;"
-                      className="w-full text-sm border rounded-md px-2.5 py-1 bg-background focus:outline-none focus:ring-1 focus:ring-primary"
-                    />
-                    <div className="flex items-center gap-2 flex-wrap">
-                      <button
-                        onClick={() => {
-                          if (confirm('Discard unsaved changes?')) discardChanges();
-                        }}
-                        className="inline-flex items-center gap-1.5 rounded-md border px-3 py-1.5 text-sm font-medium text-muted-foreground hover:text-foreground hover:bg-muted transition-colors whitespace-nowrap"
-                      >
-                        Discard changes
-                      </button>
-                      <button
-                        onClick={handleSave}
-                        disabled={saveState.status === 'saving' || !editedTitle.trim()}
-                        className={cn(
-                          'inline-flex items-center gap-1.5 rounded-md bg-primary px-3 py-1.5 text-sm font-medium text-primary-foreground hover:bg-primary/90 transition-colors whitespace-nowrap',
-                          (saveState.status === 'saving' || !editedTitle.trim()) && 'opacity-50 cursor-not-allowed',
-                        )}
-                      >
-                        <Save className="h-3.5 w-3.5" />
-                        {saveState.status === 'saving' ? 'Saving...' : 'Save new version'}
-                      </button>
-                      {saveState.status === 'saved' && (
-                        <span className="inline-flex items-center gap-1.5 rounded-md bg-green-50 border border-green-200 px-3 py-1.5 text-sm font-medium text-green-700 dark:bg-green-900/20 dark:border-green-800 dark:text-green-400">
-                          Saved as v{saveState.version}
-                        </span>
-                      )}
-                      {saveState.status === 'error' && (
-                        <span className="inline-flex items-center gap-1.5 rounded-md bg-red-50 border border-red-200 px-3 py-1.5 text-sm text-red-700 dark:bg-red-900/20 dark:border-red-800 dark:text-red-400">
-                          {saveState.message}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                )}
-              </>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// YAML editor — direct text editing of steps + transitions
-// ---------------------------------------------------------------------------
-
-function YamlEditor({ steps, transitions, onChange }: {
-  steps: WorkflowStep[];
-  transitions: WorkflowDefinition['transitions'];
-  onChange: (steps: WorkflowStep[], transitions: WorkflowDefinition['transitions']) => void;
-}) {
-  const [yamlText, setYamlText] = useState(() =>
-    yamlStringify({ steps, transitions }, { indent: 2 }),
-  );
-  const [error, setError] = useState<string | null>(null);
-
-  const applyYaml = useCallback(() => {
-    try {
-      const parsed = yamlParse(yamlText) as { steps?: WorkflowStep[]; transitions?: WorkflowDefinition['transitions'] };
-      if (!parsed?.steps || !Array.isArray(parsed.steps)) {
-        setError('YAML must contain a "steps" array');
-        return;
-      }
-      setError(null);
-      onChange(parsed.steps, parsed.transitions ?? []);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Invalid YAML');
-    }
-  }, [yamlText, onChange]);
-
-  return (
-    <div className="mt-2 space-y-2">
-      <textarea
-        value={yamlText}
-        onChange={(e) => setYamlText(e.target.value)}
-        rows={20}
-        spellCheck={false}
-        className="w-full text-[11px] font-mono bg-muted/30 rounded-lg p-4 border-0 focus:outline-none focus:ring-1 focus:ring-primary resize-y leading-relaxed"
+      {/* Editor canvas — key resets internal state when navigating to a different version */}
+      <WorkflowEditorCanvas
+        key={definition.version}
+        initialSteps={definition.steps}
+        initialTransitions={definition.transitions}
+        workflowName={decodedName}
+        yamlFields={{ ...definition, version: undefined, createdAt: undefined } as Record<string, unknown>}
+        renderSavePanel={renderSavePanel}
       />
-      {error && (
-        <p className="text-xs text-red-600 dark:text-red-400">{error}</p>
-      )}
-      <button
-        onClick={applyYaml}
-        className="text-xs font-medium text-primary hover:underline"
-      >
-        Apply YAML to diagram
-      </button>
     </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Step editor (edit mode side panel)
-// ---------------------------------------------------------------------------
-
-const STEP_TYPES = ['creation', 'review', 'decision', 'terminal'] as const;
-const FALLBACK_OPTIONS = [
-  { value: '', label: 'Default' },
-  { value: 'escalate_to_human', label: 'Escalate to human' },
-  { value: 'continue_with_flag', label: 'Continue with flag' },
-  { value: 'pause', label: 'Pause' },
-] as const;
-
-const KNOWN_MODELS = [
-  { value: '', label: 'Default' },
-  { value: 'sonnet', label: 'Sonnet' },
-  { value: 'claude-sonnet-4-6', label: 'Claude Sonnet 4.6' },
-  { value: 'claude-opus-4-6', label: 'Claude Opus 4.6' },
-  { value: 'claude-haiku-4-5-20251001', label: 'Claude Haiku 4.5' },
-] as const;
-
-const RUNTIME_OPTIONS = [
-  { value: '', label: 'None' },
-  { value: 'javascript', label: 'JavaScript' },
-  { value: 'python', label: 'Python' },
-  { value: 'r', label: 'R' },
-  { value: 'bash', label: 'Bash' },
-] as const;
-
-function toSlug(name: string): string {
-  return name.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, '');
-}
-
-function StepEditor({ step, allSteps, onChange }: { step: WorkflowStep; allSteps: WorkflowStep[]; onChange: (patch: Partial<WorkflowStep>) => void }) {
-  const isNewStep = step.id.startsWith('new-step-');
-  const { plugins } = usePlugins();
-  const { firebaseUser } = useAuth();
-  const { handle, name: workflowNameParam } = useParams<{ handle: string; name: string }>();
-  const [secretKeys, setSecretKeys] = useState<string[]>([]);
-  useEffect(() => {
-    if (handle && workflowNameParam && firebaseUser) {
-      getWorkflowSecretKeys(handle, decodeURIComponent(workflowNameParam), firebaseUser.uid)
-        .then(setSecretKeys)
-        .catch((error) => console.error('Failed to load secret keys:', error));
-    }
-  }, [handle, workflowNameParam, firebaseUser]);
-  const inlineInput = 'w-full bg-transparent border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0.5 focus:outline-none transition-colors';
-  const selectInline = 'bg-transparent text-xs text-right border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors cursor-pointer';
-  const otherSteps = allSteps.filter((s) => s.id !== step.id);
-
-  const isAgent = step.executor === 'agent' && step.type !== 'terminal';
-  const isHuman = step.executor === 'human' && step.type !== 'terminal';
-  const isReview = step.type === 'review';
-  const hasAgent = step.agent && Object.values(step.agent).some((v) => v !== undefined);
-
-  return (
-    <div className="space-y-5">
-      {/* ─── Identity — mirrors StepDetail header ─── */}
-      <div>
-        <input
-          value={step.name}
-          onChange={(e) => {
-            const patch: Partial<WorkflowStep> = { name: e.target.value };
-            // Auto-slug ID for new steps only
-            if (isNewStep) patch.id = toSlug(e.target.value) || step.id;
-            onChange(patch);
-          }}
-          className={cn(inlineInput, 'text-[15px] font-semibold text-foreground')}
-        />
-        <StepIdField currentId={step.id} onChange={(newId) => onChange({ id: newId })} />
-        <textarea
-          value={step.description ?? ''}
-          onChange={(e) => onChange({ description: e.target.value || undefined })}
-          placeholder="Add description..."
-          rows={2}
-          className={cn(inlineInput, 'mt-2 text-sm text-muted-foreground resize-y leading-relaxed placeholder:italic')}
-        />
-      </div>
-
-      {/* ─── Badges row — executor is a toggle, type + autonomy are clickable ─── */}
-      <div className="space-y-2">
-        {/* Executor — Human/Agent toggle (script steps show agent/human options too) */}
-        {step.type !== 'terminal' && step.executor !== 'script' && (
-          <div className="flex gap-1 p-0.5 rounded-lg bg-muted">
-            {(['human', 'agent'] as const).map((ex) => {
-              const Icon = ex === 'human' ? User : Bot;
-              const activeColors: Record<string, string> = {
-                human: 'bg-blue-500 text-white shadow-sm',
-                agent: 'bg-violet-500 text-white shadow-sm',
-              };
-              return (
-                <button
-                  key={ex}
-                  onClick={() => onChange({
-                    executor: ex,
-                    // Default plugin when switching to agent (prevents "plugin not registered" errors)
-                    ...(ex === 'agent' && !step.plugin ? { plugin: 'opencode-agent' } : {}),
-                  })}
-                  className={cn(
-                    'flex-1 flex items-center justify-center gap-1.5 rounded-md px-2 py-2 text-xs font-medium capitalize transition-all',
-                    step.executor === ex ? activeColors[ex] : 'text-muted-foreground hover:text-foreground',
-                  )}
-                >
-                  <Icon className="h-3.5 w-3.5" />
-                  {ex}
-                </button>
-              );
-            })}
-          </div>
-        )}
-
-        {/* Autonomy — the #1 tunable */}
-        {isAgent && (
-          <div className="flex flex-col gap-0.5">
-            {AUTONOMY_LEVELS.map((level) => (
-              <button
-                key={level.value}
-                onClick={() => onChange({ autonomyLevel: level.value })}
-                className={cn(
-                  'rounded-md px-3 py-1.5 text-xs text-left transition-all border',
-                  step.autonomyLevel === level.value
-                    ? 'border-violet-400 bg-violet-50 text-violet-700 font-medium dark:bg-violet-900/20 dark:text-violet-300'
-                    : 'border-transparent text-muted-foreground hover:text-foreground hover:bg-muted/50',
-                )}
-              >
-                {level.label}
-              </button>
-            ))}
-          </div>
-        )}
-      </div>
-
-      {/* ─── Core config — mirrors StepDetail fields, but editable ─── */}
-      <div className="space-y-1.5">
-        {isAgent && (
-          <div className="flex items-baseline justify-between gap-3">
-            <span className="text-xs text-muted-foreground shrink-0">Plugin</span>
-            <select
-              value={step.plugin ?? ''}
-              onChange={(e) => onChange({ plugin: e.target.value || undefined })}
-              className={selectInline}
-            >
-              <option value="">None</option>
-              {plugins.map((p) => (
-                <option key={p.name} value={p.name}>{p.name}</option>
-              ))}
-              {step.plugin && !plugins.some((p) => p.name === step.plugin) && (
-                <option value={step.plugin}>{step.plugin}</option>
-              )}
-            </select>
-          </div>
-        )}
-        {step.executor === 'script' && step.type !== 'terminal' && (
-          <EditableField label="Plugin" value={step.plugin ?? ''} mono placeholder="e.g. script-container"
-            onChange={(v) => onChange({ plugin: v || undefined })} />
-        )}
-        {isHuman && (
-          <EditableField label="Roles" value={step.allowedRoles?.join(', ') ?? ''} placeholder="e.g. qa-lead, analyst"
-            onChange={(v) => onChange({ allowedRoles: v ? v.split(',').map((r) => r.trim()).filter(Boolean) : undefined })} />
-        )}
-      </div>
-
-      {/* ─── Agent section — mirrors StepDetail "Agent" section ─── */}
-      {isAgent && (
-        <Section title="Agent">
-          <div className="flex items-baseline justify-between gap-3">
-            <span className="text-xs text-muted-foreground shrink-0">Model</span>
-            <select
-              value={step.agent?.model ?? ''}
-              onChange={(e) => onChange({ agent: { ...step.agent, model: e.target.value || undefined } })}
-              className={selectInline}
-            >
-              {KNOWN_MODELS.map((m) => (
-                <option key={m.value} value={m.value}>{m.label}</option>
-              ))}
-              {step.agent?.model && !KNOWN_MODELS.some((m) => m.value === step.agent?.model) && (
-                <option value={step.agent.model}>{step.agent.model}</option>
-              )}
-            </select>
-          </div>
-          <EditableField label="Skill" value={step.agent?.skill ?? ''} mono placeholder="skill-name"
-            onChange={(v) => onChange({ agent: { ...step.agent, skill: v || undefined } })} />
-          <EditableField label="Timeout" value={step.agent?.timeoutMinutes !== undefined ? `${step.agent.timeoutMinutes}` : ''} placeholder="30"
-            onChange={(v) => onChange({ agent: { ...step.agent, timeoutMinutes: v ? Number(v) : undefined } })} suffix="min" />
-          <EditableField label="Confidence" value={step.agent?.confidenceThreshold !== undefined ? `${step.agent.confidenceThreshold}` : ''} placeholder="0.85"
-            onChange={(v) => onChange({ agent: { ...step.agent, confidenceThreshold: v ? Number(v) : undefined } })} />
-          <div className="flex items-baseline justify-between gap-3">
-            <span className="text-xs text-muted-foreground shrink-0">Fallback</span>
-            <select
-              value={step.agent?.fallbackBehavior ?? ''}
-              onChange={(e) => onChange({ agent: { ...step.agent, fallbackBehavior: (e.target.value || undefined) as 'escalate_to_human' | 'continue_with_flag' | 'pause' | undefined } })}
-              className="bg-transparent text-xs text-right border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors cursor-pointer"
-            >
-              {FALLBACK_OPTIONS.map((opt) => (
-                <option key={opt.value} value={opt.value}>{opt.label}</option>
-              ))}
-            </select>
-          </div>
-          {(hasAgent || true) && (
-            <div className="mt-2">
-              <p className="text-[11px] text-muted-foreground mb-1">Prompt</p>
-              <textarea
-                value={step.agent?.prompt ?? ''}
-                onChange={(e) => onChange({ agent: { ...step.agent, prompt: e.target.value || undefined } })}
-                rows={3}
-                placeholder="Instructions for the agent..."
-                className="w-full text-xs bg-muted/50 rounded-md p-2.5 leading-relaxed border-0 focus:outline-none focus:ring-1 focus:ring-primary resize-y"
-              />
-            </div>
-          )}
-        </Section>
-      )}
-
-      {/* ─── Review section ─── */}
-      {isReview && (
-        <Section title="Review">
-          <div className="flex gap-1 mb-2">
-            {(['human', 'agent', 'none'] as const).map((rt) => (
-              <button
-                key={rt}
-                onClick={() => onChange({ review: { ...step.review, type: rt } })}
-                className={cn(
-                  'flex-1 rounded-md py-1 text-[11px] font-medium capitalize transition-all border',
-                  step.review?.type === rt
-                    ? 'border-primary bg-primary/5 text-primary'
-                    : 'border-transparent bg-muted/50 text-muted-foreground hover:text-foreground',
-                )}
-              >
-                {rt}
-              </button>
-            ))}
-          </div>
-          <EditableField label="Plugin" value={step.review?.plugin ?? ''} mono placeholder="review-plugin"
-            onChange={(v) => onChange({ review: { ...step.review, plugin: v || undefined } })} />
-          <EditableField label="Max iterations" value={step.review?.maxIterations !== undefined ? `${step.review.maxIterations}` : ''} placeholder="3"
-            onChange={(v) => onChange({ review: { ...step.review, maxIterations: v ? Number(v) : undefined } })} />
-          <EditableField label="Time box" value={step.review?.timeBoxDays !== undefined ? `${step.review.timeBoxDays}` : ''} placeholder="5"
-            onChange={(v) => onChange({ review: { ...step.review, timeBoxDays: v ? Number(v) : undefined } })} suffix="days" />
-        </Section>
-      )}
-
-      {/* ─── Verdicts (review steps) ─── */}
-      {isReview && (
-        <Section title="Verdicts">
-          <div className="space-y-1.5">
-            {Object.entries(step.verdicts ?? {}).map(([verdictName, verdict]) => (
-              <div key={verdictName} className="flex items-center gap-1.5">
-                <input
-                  value={verdictName}
-                  onChange={(e) => {
-                    const newVerdicts: Record<string, { target: string }> = {};
-                    for (const [k, v] of Object.entries(step.verdicts ?? {})) {
-                      newVerdicts[k === verdictName ? e.target.value : k] = v;
-                    }
-                    onChange({ verdicts: newVerdicts });
-                  }}
-                  className="bg-transparent text-xs font-medium border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors w-20"
-                />
-                <span className="text-xs text-muted-foreground">→</span>
-                <select
-                  value={verdict.target}
-                  onChange={(e) => {
-                    const newVerdicts = { ...step.verdicts, [verdictName]: { ...verdict, target: e.target.value } };
-                    onChange({ verdicts: newVerdicts });
-                  }}
-                  className={cn(selectInline, 'flex-1')}
-                >
-                  <option value="">Select step...</option>
-                  {otherSteps.map((s) => (
-                    <option key={s.id} value={s.id}>{s.name} ({s.id})</option>
-                  ))}
-                </select>
-                <button
-                  onClick={() => {
-                    const newVerdicts = { ...step.verdicts };
-                    delete newVerdicts[verdictName];
-                    onChange({ verdicts: Object.keys(newVerdicts).length > 0 ? newVerdicts : undefined });
-                  }}
-                  className="text-[10px] text-muted-foreground/40 hover:text-red-500 transition-colors"
-                >
-                  ×
-                </button>
-              </div>
-            ))}
-            <button
-              onClick={() => {
-                onChange({ verdicts: { ...step.verdicts, ['new-verdict']: { target: '' } } });
-              }}
-              className="text-[11px] text-muted-foreground hover:text-foreground transition-colors"
-            >
-              + Add verdict
-            </button>
-          </div>
-        </Section>
-      )}
-
-      {/* ─── Runtime — for script/agent steps with runtime config ─── */}
-      {(step.executor === 'script' || step.executor === 'agent') && step.type !== 'terminal' && (
-        <Section title="Runtime">
-          <div className="flex items-baseline justify-between gap-3">
-            <span className="text-xs text-muted-foreground shrink-0">Runtime</span>
-            <select
-              value={step.agent?.runtime ?? ''}
-              onChange={(e) => onChange({ agent: { ...step.agent, runtime: (e.target.value || undefined) as 'javascript' | 'python' | 'r' | 'bash' | undefined } })}
-              className={selectInline}
-            >
-              {RUNTIME_OPTIONS.map((o) => (
-                <option key={o.value} value={o.value}>{o.label}</option>
-              ))}
-            </select>
-          </div>
-          <EditableField label="Command" value={step.agent?.command ?? ''} mono placeholder="run.sh"
-            onChange={(v) => onChange({ agent: { ...step.agent, command: v || undefined } })} />
-          <EditableField label="Image" value={step.agent?.image ?? ''} mono placeholder="docker-image:tag"
-            onChange={(v) => onChange({ agent: { ...step.agent, image: v || undefined } })} />
-          <div>
-            <p className="text-[11px] text-muted-foreground mb-1">Inline script</p>
-            <textarea
-              value={step.agent?.inlineScript ?? ''}
-              onChange={(e) => onChange({ agent: { ...step.agent, inlineScript: e.target.value || undefined } })}
-              rows={3}
-              placeholder="Script code..."
-              className="w-full text-[11px] font-mono bg-muted/50 rounded-md p-2 leading-relaxed border-0 focus:outline-none focus:ring-1 focus:ring-primary resize-y"
-            />
-          </div>
-        </Section>
-      )}
-
-      {/* ─── Environment variables ─── */}
-      {step.type !== 'terminal' && (
-        <Section title="Environment">
-          {Object.entries(step.env ?? {}).map(([key, val]) => (
-            <div key={key} className="flex items-baseline gap-1.5">
-              <input
-                value={key}
-                onChange={(e) => {
-                  const newEnv: Record<string, string> = {};
-                  for (const [k, v] of Object.entries(step.env ?? {})) {
-                    newEnv[k === key ? e.target.value : k] = v;
-                  }
-                  onChange({ env: newEnv });
-                }}
-                className="bg-transparent text-xs font-mono text-muted-foreground border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors w-24"
-              />
-              <span className="text-xs text-muted-foreground">=</span>
-              <div className="relative flex-1 group">
-                <input
-                  value={val}
-                  onChange={(e) => {
-                    const newEnv = { ...step.env, [key]: e.target.value };
-                    onChange({ env: newEnv });
-                  }}
-                  className="bg-transparent text-xs font-mono border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors w-full"
-                />
-                {secretKeys.length > 0 && !val.startsWith('{{') && (
-                  <select
-                    value=""
-                    onChange={(e) => {
-                      if (e.target.value) {
-                        const newEnv = { ...step.env, [key]: `{{${e.target.value}}}` };
-                        onChange({ env: newEnv });
-                      }
-                    }}
-                    className="absolute right-0 top-0 h-full opacity-0 group-hover:opacity-100 focus:opacity-100 bg-transparent text-xs cursor-pointer transition-opacity w-5"
-                    title="Insert secret reference"
-                  >
-                    <option value="">🔑</option>
-                    {secretKeys.map((sk) => (
-                      <option key={sk} value={sk}>{sk}</option>
-                    ))}
-                  </select>
-                )}
-              </div>
-              <button
-                onClick={() => {
-                  const newEnv = { ...step.env };
-                  delete newEnv[key];
-                  onChange({ env: Object.keys(newEnv).length > 0 ? newEnv : undefined });
-                }}
-                className="text-[10px] text-muted-foreground/40 hover:text-red-500 transition-colors"
-              >
-                ×
-              </button>
-            </div>
-          ))}
-          <button
-            onClick={() => {
-              onChange({ env: { ...step.env, NEW_VAR: '' } });
-            }}
-            className="text-[11px] text-muted-foreground hover:text-foreground transition-colors"
-          >
-            + Add variable
-          </button>
-        </Section>
-      )}
-
-      {/* ─── Step definition (collapsed — rarely changed) ─── */}
-      <details className="group">
-        <summary className="text-[11px] font-semibold uppercase tracking-widest text-muted-foreground/40 cursor-pointer hover:text-muted-foreground transition-colors select-none">
-          Step definition
-        </summary>
-        <div className="mt-2 space-y-2.5">
-          <div className="flex gap-1">
-            {STEP_TYPES.map((t) => (
-              <button
-                key={t}
-                onClick={() => onChange({ type: t })}
-                className={cn(
-                  'flex-1 rounded-md py-1 text-[11px] font-medium capitalize transition-all border',
-                  step.type === t
-                    ? 'border-primary bg-primary/5 text-primary'
-                    : 'border-transparent bg-muted/50 text-muted-foreground hover:text-foreground',
-                )}
-              >
-                {t}
-              </button>
-            ))}
-          </div>
-          {/* Step ID is now editable in the identity section above */}
-          {step.type !== 'terminal' && (
-            <div className="flex items-center justify-between pt-1">
-              <span className="text-xs text-muted-foreground">Automated step</span>
-              <button
-                onClick={() => onChange({ executor: step.executor === 'script' ? 'human' : 'script' })}
-                className={cn(
-                  'relative inline-flex h-5 w-9 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors',
-                  step.executor === 'script' ? 'bg-amber-500' : 'bg-muted',
-                )}
-              >
-                <span className={cn(
-                  'pointer-events-none inline-block h-4 w-4 rounded-full bg-white shadow-sm transition-transform',
-                  step.executor === 'script' ? 'translate-x-4' : 'translate-x-0',
-                )} />
-              </button>
-            </div>
-          )}
-        </div>
-      </details>
-    </div>
-  );
-}
-
-function StepIdField({ currentId, onChange }: { currentId: string; onChange: (newId: string) => void }) {
-  const [draft, setDraft] = useState(currentId);
-  const [dirty, setDirty] = useState(false);
-  const prevIdRef = useRef(currentId);
-
-  // Sync draft immediately when external id changes (e.g. auto-slug from name)
-  if (currentId !== prevIdRef.current && !dirty) {
-    prevIdRef.current = currentId;
-    // Direct state set during render — React handles this correctly
-    setDraft(currentId);
-  }
-  prevIdRef.current = currentId;
-
-  const commit = useCallback(() => {
-    const slug = draft.toLowerCase().replace(/[^a-z0-9-]/g, '-').replace(/-+/g, '-').replace(/^-|-$/g, '');
-    if (slug && slug !== currentId) {
-      onChange(slug);
-    }
-    setDraft(slug || currentId);
-    setDirty(false);
-  }, [draft, currentId, onChange]);
-
-  return (
-    <input
-      value={draft}
-      onChange={(e) => { setDraft(e.target.value); setDirty(true); }}
-      onBlur={commit}
-      onKeyDown={(e) => { if (e.key === 'Enter') commit(); }}
-      placeholder="step-id"
-      className="w-full bg-transparent border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0.5 focus:outline-none transition-colors font-mono text-xs text-muted-foreground mt-0.5"
-    />
-  );
-}
-
-function EditableField({ label, value, onChange, mono, placeholder, suffix }: {
-  label: string; value: string; onChange: (v: string) => void;
-  mono?: boolean; placeholder?: string; suffix?: string;
-}) {
-  return (
-    <div className="flex items-baseline justify-between gap-3">
-      <span className="text-xs text-muted-foreground shrink-0">{label}</span>
-      <div className="flex items-baseline gap-1">
-        <input
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          placeholder={placeholder}
-          className={cn(
-            'bg-transparent text-xs text-right border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors w-28',
-            mono && 'font-mono',
-            !value && 'placeholder:text-muted-foreground/40 placeholder:italic',
-          )}
-        />
-        {suffix && <span className="text-[10px] text-muted-foreground">{suffix}</span>}
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Step detail (read-only side panel)
-// ---------------------------------------------------------------------------
-
-const FALLBACK_LABELS: Record<string, string> = {
-  escalate_to_human: 'Escalate to human',
-  continue_with_flag: 'Continue with flag',
-  pause: 'Pause',
-};
-
-function StepDetail({ step }: { step: WorkflowStep }) {
-  const hasAgent = step.agent && Object.values(step.agent).some((v) => v !== undefined);
-  const hasReview = step.review && Object.values(step.review).some((v) => v !== undefined);
-  const hasVerdicts = step.verdicts && Object.keys(step.verdicts).length > 0;
-
-  return (
-    <div className="space-y-5">
-      {/* Identity */}
-      <div>
-        <p className="text-[15px] font-semibold">{step.name}</p>
-        <p className="font-mono text-xs text-muted-foreground mt-0.5">{step.id}</p>
-        {step.description && (
-          <p className="text-sm text-muted-foreground mt-2 leading-relaxed">{step.description}</p>
-        )}
-      </div>
-
-      {/* Badges */}
-      <div className="flex flex-wrap gap-1.5">
-        <StepTypeBadge type={step.type} />
-        {step.type !== 'terminal' && <ExecutorBadge executor={step.executor} />}
-        {step.type !== 'terminal' && step.autonomyLevel && (
-          <span className="rounded-full bg-violet-50 px-2 py-0.5 text-xs font-medium text-violet-700 dark:bg-violet-900/30 dark:text-violet-300">
-            {step.autonomyLevel}
-          </span>
-        )}
-      </div>
-
-      {/* Core config */}
-      <div className="space-y-1.5">
-        {step.plugin && <Field label="Plugin" value={step.plugin} mono />}
-        {step.allowedRoles && step.allowedRoles.length > 0 && (
-          <Field label="Roles" value={step.allowedRoles.join(', ')} />
-        )}
-      </div>
-
-      {/* Agent config */}
-      {hasAgent && step.executor === 'agent' && (
-        <Section title="Agent">
-          {step.agent!.model && <Field label="Model" value={step.agent!.model} mono />}
-          {step.agent!.skill && <Field label="Skill" value={step.agent!.skill} mono />}
-          {step.agent!.timeoutMinutes !== undefined && <Field label="Timeout" value={`${step.agent!.timeoutMinutes} min`} />}
-          {step.agent!.confidenceThreshold !== undefined && <Field label="Confidence" value={`${step.agent!.confidenceThreshold}`} />}
-          {step.agent!.fallbackBehavior && (
-            <Field label="Fallback" value={FALLBACK_LABELS[step.agent!.fallbackBehavior] ?? step.agent!.fallbackBehavior} />
-          )}
-          {step.agent!.prompt && (
-            <div className="mt-2">
-              <p className="text-[11px] text-muted-foreground mb-1">Prompt</p>
-              <p className="w-full text-xs bg-muted/50 rounded-md p-2.5 whitespace-pre-line break-words leading-relaxed text-justify">{step.agent!.prompt}</p>
-            </div>
-          )}
-        </Section>
-      )}
-
-      {/* Review config */}
-      {hasReview && (
-        <Section title="Review">
-          {step.review!.type && <Field label="Reviewer" value={step.review!.type} />}
-          {step.review!.plugin && <Field label="Plugin" value={step.review!.plugin} mono />}
-          {step.review!.maxIterations !== undefined && <Field label="Max iterations" value={String(step.review!.maxIterations)} />}
-          {step.review!.timeBoxDays !== undefined && <Field label="Time box" value={`${step.review!.timeBoxDays} days`} />}
-        </Section>
-      )}
-
-      {/* Verdicts */}
-      {hasVerdicts && (
-        <Section title="Verdicts">
-          <div className="space-y-1.5">
-            {Object.entries(step.verdicts!).map(([name, verdict]) => (
-              <div key={name} className="flex items-center gap-2 text-xs">
-                <span className="rounded bg-muted px-1.5 py-0.5 font-medium">{name}</span>
-                <span className="text-muted-foreground">→</span>
-                <span className="font-mono text-muted-foreground">{verdict.target}</span>
-              </div>
-            ))}
-          </div>
-        </Section>
-      )}
-
-      {/* Env vars */}
-      {step.env && Object.keys(step.env).length > 0 && (
-        <Section title="Environment">
-          {Object.entries(step.env).map(([key, val]) => (
-            <Field key={key} label={key} value={val} mono />
-          ))}
-        </Section>
-      )}
-
-      {/* Script & runtime config, step params */}
-      {(step.agent?.command || step.agent?.inlineScript || step.agent?.runtime || step.agent?.image || step.stepParams) && (
-        <Section title="Runtime">
-          {step.agent?.runtime && <Field label="Runtime" value={step.agent.runtime} />}
-          {step.agent?.command && <Field label="Command" value={step.agent.command} mono />}
-          {step.agent?.inlineScript && (
-            <details>
-              <summary className="text-[11px] text-muted-foreground cursor-pointer hover:text-foreground transition-colors">Inline script</summary>
-              <pre className="mt-1 text-[10px] bg-muted/50 rounded-md p-2 overflow-x-auto font-mono">{step.agent.inlineScript}</pre>
-            </details>
-          )}
-          {step.agent?.image && <Field label="Image" value={step.agent.image} mono />}
-          {step.agent?.repo && <Field label="Repo" value={step.agent.repo} mono />}
-          {step.agent?.commit && <Field label="Commit" value={step.agent.commit} mono />}
-          {step.stepParams && Object.keys(step.stepParams).length > 0 && (
-            <div>
-              <p className="text-[11px] text-muted-foreground mb-1">Parameters</p>
-              <pre className="text-[10px] bg-muted/50 rounded-md p-2 overflow-x-auto font-mono">{JSON.stringify(step.stepParams, null, 2)}</pre>
-            </div>
-          )}
-        </Section>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Shared UI helpers
-// ---------------------------------------------------------------------------
-
-function Section({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <div className="space-y-1.5">
-      <p className="text-[11px] font-semibold uppercase tracking-widest text-muted-foreground/60">{title}</p>
-      {children}
-    </div>
-  );
-}
-
-function Field({ label, value, mono }: { label: string; value: string; mono?: boolean }) {
-  return (
-    <div className="flex items-baseline justify-between gap-3">
-      <span className="text-xs text-muted-foreground shrink-0">{label}</span>
-      <span className={cn('text-xs text-right truncate', mono && 'font-mono')}>{value}</span>
-    </div>
-  );
-}
-
-function ExecutorBadge({ executor }: { executor: string }) {
-  const colors: Record<string, string> = {
-    human: 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-400',
-    agent: 'bg-violet-100 text-violet-700 dark:bg-violet-900/30 dark:text-violet-400',
-    script: 'bg-amber-100 text-amber-700 dark:bg-amber-900/30 dark:text-amber-400',
-  };
-  return (
-    <span className={cn('rounded-full px-2 py-0.5 text-xs font-medium capitalize', colors[executor] ?? 'bg-muted text-muted-foreground')}>
-      {executor}
-    </span>
-  );
-}
-
-const STEP_TYPE_LABELS: Record<string, string> = {
-  creation: 'Task',
-  review: 'Review',
-  decision: 'Decision',
-  terminal: 'End',
-};
-
-function StepTypeBadge({ type }: { type: string }) {
-  const colors: Record<string, string> = {
-    creation: 'bg-blue-50 text-blue-600 dark:bg-blue-900/20 dark:text-blue-400',
-    review: 'bg-amber-50 text-amber-600 dark:bg-amber-900/20 dark:text-amber-400',
-    decision: 'bg-purple-50 text-purple-600 dark:bg-purple-900/20 dark:text-purple-400',
-    terminal: 'bg-slate-100 text-slate-500 dark:bg-slate-800 dark:text-slate-400',
-  };
-  return (
-    <span className={cn('rounded-full px-2 py-0.5 text-xs font-medium', colors[type] ?? 'bg-muted text-muted-foreground')}>
-      {STEP_TYPE_LABELS[type] ?? type}
-    </span>
   );
 }
